Rename all 'food ticket' values to 'food event'
